requirements.txt
streamlit==1.35.0
pandas==2.2.3
matplotlib==3.8.4
cassandra-driver==3.29.1
